updated integer foreign keys for groups, players and games to bigints

+++ schema.sql
@@ -17,8 +17,8 @@ CREATE TABLE games (
 
 CREATE TABLE games_and_groups (
     id SERIAL PRIMARY KEY,
-    game_id INTEGER NOT NULL,
-    group_id INTEGER NOT NULL,
+    game_id BIGINT NOT NULL,
+    group_id BIGINT NOT NULL,
     FOREIGN KEY(game_id) REFERENCES games(game_id),
     FOREIGN KEY(group_id) REFERENCES groups(group_id)
 );
@@ -28,14 +28,14 @@ CREATE TABLE players (
     password_hash VARCHAR(100) NOT NULL,
     player_name VARCHAR(20) NOT NULL,
     picked_name VARCHAR(20) DEFAULT NULL,
-    group_id INTEGER DEFAULT NULL,
+    group_id BIGINT DEFAULT NULL,
     FOREIGN KEY(group_id) REFERENCES groups(group_id)
 );
 
 CREATE TABLE games_and_players (
     id SERIAL PRIMARY KEY,
-    game_id INTEGER NOT NULL,
-    player_id INTEGER NOT NULL,
+    game_id BIGINT NOT NULL,
+    player_id BIGINT NOT NULL,
     FOREIGN KEY(game_id) REFERENCES games(game_id),
     FOREIGN KEY(player_id) REFERENCES players(player_id)
 );
